Monthly and Biweekly Integration Added

+++ database/migrations/2021_10_28_150406_create_payments_table.php
@@ -18,7 +18,7 @@ class CreatePaymentsTable extends Migration
             $table->timestamps();
             $table->bigInteger('user')->unsigned();
             $table->foreign('user')->references('id')->on('users')->onDelete('cascade');
-            $table->bigInteger('transaction')->unsigned()->nullable();
+            $table->double('transaction')->unsigned()->nullable();
             $table->foreign('transaction')->references('id')->on('transactions')->onDelete('cascade');
             $table->integer('amount')->unsigned();
             $table->boolean('completed');
